Changed VoogaDaemon, VoogaServer, and voogaGames.xml
VoogaServer:  Was some repetitive and difficult to read code so I made a map of VoogaDaemons that can easily be iterated to get the port number or other information.  Also I pulled the big chunk of code out of the main method so that it won't get too confusing and busy if we decide to add to the main method.  

VoogaDaemon now has public int fields gamePort and chatPort that are referenced when you call getPort on VoogaServer instead of the XML file.

voogaGames.xml:  changed "port" to "gamePort".  More consistent with "chatPort".